requirements.txt
# === Core Deep Learning ===
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# === Hugging Face Ecosystem ===
transformers==4.44.2
datasets>=2.19.0
accelerate==0.34.2
evaluate>=0.4.0
huggingface-hub>=0.25.0  

# === Image Processing & Augmentation ===
Pillow>=10.0.0
albumentations>=1.4.4
opencv-python-headless>=4.9.0  


# === Data Science Utilities ===
numpy>=1.26.4
pandas>=2.2.0
scikit-learn>=1.4.2  
tqdm>=4.66.2

# === Logging / Visualization ===
rich>=13.7.1
ipywidgets>=8.1.1
matplotlib>=3.8.0

# === Misc ===
python-dotenv>=1.0.1
timm>=1.0.8  

# === Safety & File Handling ===
pathlib>=1.0.1  
